[zen-observable] re-export types from the entrypoint
Instead of declaring types in the 'esm' definitions, keep them in 'index' and re-export them in 'esm'.

This commit also moves the 'extras' types in a separate module, to be closer to the upstream package.
